Code cleanup to avoid compiler warnings.
git-svn-id: https://core.forge.funambol.org/svn/core/top-level/trunk/3x/client-api/native@24517 e8e8ed6c-164c-0410-afcf-9e9a7c7d8c10

+++ src/c++/symbian/push/FSocket.cpp
@@ -235,7 +235,8 @@ int32_t FSocket::readBuffer(int8_t* buffer, int32_t maxLen)
             return msgLen;
         }
         else {
-            errorMsg.sprintf("FSocket: error reading, message too big (%d > %d) -> rejected.", msgLen, maxLen);
+            errorMsg.sprintf("FSocket: error reading, message too big (%d > %d) -> rejected.",
+                             (int)msgLen, (int)maxLen);
             goto error;
         }
     }
